fatal-signal: add graceful shutdown functionality
Add a graceful shutdown functionality based on the
following functions:

fatal_signal_enable_graceful_shutdown(): Overrides the
SIGTERM handler to avoid exiting and instead store in a
variable if the signal was received.

fatal_signal_should_graceful_exit(): Returns true
if the SIGTERM signal was received. It should be used
on the poll main loops to break the loop, clean
the resources and exit.

Signed-off-by: Antonio Ojea <aojea@redhat.com>

asd